forensics/protochess/setup/pyproject.toml
[tool.poetry]
name = "grpc-enc"
version = "0.1.0"
description = ""
authors = ["rok0s <info@cybermouflons.com>"]

[tool.poetry.dependencies]
python = "^3.7"
grpcio-tools = "^1.36.1"
pycryptodome = "^3.10.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
